# smart-home-msa-solution/smart-home-app/device-manager/swagger.yml
openapi: 3.0.0
info:
  title: Device Manager API
  description: >-
    API for managing devices, scenarios, and telemetry data in the Smart Home
    system.
  version: 1.0.0
...
components:
...
  responses:
...
    DeviceDetail:
      description: A detailed device object
      content:
        application/json:
          example:
            id: 3
            serial_number: 1-REWQ54326
            device_type_id: 2
            house_id: 4
            is_on: true
          schema:
            type: object
            properties:
              id:
                type: integer
              serial_number:
                type: string
              device_type_id:
                type: integer
              house_id:
                type: integer
              is_on:
                type: boolean
    DeviceDeleted:
      description: Device deleted successfully
      content:
        application/json:
          example:
            message: device_id = 1
...
    TelemetryDetail:
      description: Telemetry detailed data
      content:
        application/json:
          example:
            device_id: 1
            data:
              serial_number: ABCDE12345
              device_type_id: 1
              house_id: 1
              is_on: true
          schema:
            type: object
            properties:
              device_id:
                type: integer
              data:
                type: object
                additionalProperties: true
...
paths:
...
  /device/{deviceId}:
    get:
      summary: Get device details
      tags:
        - Device Controller
      operationId: getDevice
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/DeviceDetail'
        '404':
          description: Device not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a device
      tags:
        - Device Controller
      operationId: deleteDevice
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          $ref: '#/components/responses/DeviceDeleted'
        '404':
          description: Device not found
        '500':
          description: Internal server error
...
  /telemetry/{telemetryId}:
    get:
      x-kusk:
        upstream:
          service:
            name: telemetry
            port: 3001
            namespace: default
          rewrite:
            pattern: "^/"
            substitution: "/api/"
      summary: Get device details
      tags:
        - Device Controller
      operationId: getDevice
      parameters:
        - name: telemetryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/TelemetryDetail'
        '404':
          description: Device not found
        '500':
          description: Internal server error
...